misc/HtmlTestSuite/results/test-0501.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Grid: automatic minimum for items with instrinsic aspect ratios</title>
<link href="mailto:svillar@igalia.com" rel="author" title="Sergio Villar"/>
<link href="https://drafts.csswg.org/css-grid/#min-size-auto" rel="help"/>
<meta content="Check how min-width|height:auto is computed for items with intrinsic aspect ratios." name="assert"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<link href="/css/support/width-keyword-classes.css" rel="stylesheet"/>

<style>
.container {
    border: 5px solid orange;
    display: grid;
}
</style>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

</head><body onload="checkLayout('.container')">
<div id="log"/>

<div class="container min-content" data-expected-height="35" data-expected-width="60">
  <img data-expected-height="25" data-expected-width="50" src="/images/green-100x50.png" style="height: 25px;"/>
</div>

<div class="container min-content" data-expected-height="80" data-expected-width="150">
  <img data-expected-height="70" data-expected-width="140" src="/images/green-100x50.png" style="min-height: 70px;"/>
</div>

<div class="container min-content" data-expected-height="50" data-expected-width="90">
  <img data-expected-height="40" data-expected-width="80" src="/images/green-100x50.png" style="max-height: 40px;"/>
</div>

<div class="container min-content" data-expected-height="40" data-expected-width="70">
  <img data-expected-height="30" data-expected-width="60" src="/images/green-100x50.png" style="height: 25px; min-height: 30px;"/>
</div>

<div class="container min-content" data-expected-height="20" data-expected-width="30">
  <img data-expected-height="10" data-expected-width="20" src="/images/green-100x50.png" style="height: 25px; max-height: 10px;"/>
</div>

<div class="container min-content" data-expected-height="35" data-expected-width="50">
  <img data-expected-height="25" data-expected-width="40" src="/images/green-100x50.png" style="height: 25px; max-width: 40px;"/>
</div>

<div class="container min-content" data-expected-height="40" data-expected-width="60">
  <img data-expected-height="30" data-expected-width="50" src="/images/green-100x50.png" style="height: 25px; min-height: 30px; max-width: 50px;"/>
</div>

<div class="container min-content" data-expected-height="35" data-expected-width="60">
  <img data-expected-height="25" data-expected-width="50" src="/images/green-100x50.png" style="width: 50px;"/>
</div>

<div class="container min-content" data-expected-height="40" data-expected-width="70">
  <img data-expected-height="30" data-expected-width="60" src="/images/green-100x50.png" style="max-width: 60px;"/>
</div>

<div class="container min-content" data-expected-height="70" data-expected-width="130">
  <img data-expected-height="60" data-expected-width="120" src="/images/green-100x50.png" style="min-width: 120px;"/>
</div>

<div class="container min-content" data-expected-height="50" data-expected-width="90">
  <img data-expected-height="40" data-expected-width="80" src="/images/green-100x50.png" style="width: 60px; min-width: 80px;"/>
</div>

<div class="container min-content" data-expected-height="22" data-expected-width="34">
  <img data-expected-height="12" data-expected-width="24" src="/images/green-100x50.png" style="width: 24px; max-width: 40px;"/>
</div>

<div class="container min-content" data-expected-height="35" data-expected-width="130">
  <img data-expected-height="25" data-expected-width="120" src="/images/green-100x50.png" style="width: 120px; max-height: 25px;"/>
</div>

<div class="container min-content" data-expected-height="110" data-expected-width="35">
  <img data-expected-height="100" data-expected-width="25" src="/images/green-100x50.png" style="width: 25px; min-height: 100px;"/>
</div>
</body>
</html>
